Use explicit cast to deal with float to double promotion from va_arg in cds_create_float_array

+++ contiguous_array/private/array_variadic.c
@@ -193,7 +193,7 @@ struct cds_array* cds_create_float_array(const size_t values_count, ...){
     for (size_t i = 0; i < values_count; ++i, ++data)
     // The float values are promoted to double when retrieved from va_arg.
     // See: https://stackoverflow.com/a/11270603/12483740
-        *data = va_arg(args, double);
+        *data = (float)va_arg(args, double);
     va_end(args);
     return array;
 }
